pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-bridge"
version = "0.1.0"
description = "Skill-Bridge Career Navigator — AI-powered skill gap analysis and learning roadmap generator."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Devansh"},
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "networkx>=3.2",
    "google-genai>=1.0.0",
    "PyMuPDF>=1.24.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
